Refactor styles and components across the application
- Replaced GlobalStyle with global.css for consistent styling.
- Updated BuySuccess page to import ButtonShop from ButtonComponents.
- Created FinalizarCompraPage with form for personal, delivery, and payment information.
- Added styles for FinalizarCompraPage.
- Introduced HomePage.css for styling HomePage components.
- Removed styled-components from HomePage and replaced with CSS classes.
- Updated Layout component to use CSS for styling instead of styled-components.
- Cleaned up unused styles in ProductListingPage and ProductViewPage.
- Enhanced global styles with a comprehensive color palette and typography settings.
