+++ src/components/Buttons/Buttons.css
@@ -1,0 +1,91 @@
+/* Button Styles */
+
+.btn-primary {
+  cursor: pointer;
+  width: 114px;
+  height: 40px;
+  border-radius: 8px;
+  background-color: var(--primary-color);
+  font-weight: 700;
+  font-size: 0.85rem;
+  letter-spacing: 0.75px;
+  line-height: 22px;
+  transition: all 0.3s ease-in-out;
+  color: var(--light-gray-3);
+  border: none;
+}
+
+.btn-primary:hover {
+  background-color: var(--tertiary-color);
+}
+
+.btn-primary.desktop {
+  width: 220px;
+  height: 48px;
+}
+
+.btn-primary.btn-add-to-cart {
+  width: 100%;
+}
+
+.btn-primary.btn-login {
+  width: 100%;
+  height: 48px;
+}
+
+@media (max-width: 1400px) {
+  .btn-primary.desktop {
+    width: 200px;
+    height: 40px;
+  }
+}
+
+@media (max-width: 460px) {
+  .btn-primary.mobile {
+    display: block;
+  }
+  
+  .btn-primary:not(.mobile) {
+    display: none;
+  }
+  
+  .btn-primary.btn-add-to-cart {
+    width: 130px;
+    height: 22px;
+    font-weight: 500;
+    font-size: 0.6rem;
+    margin-top: 10px;
+  }
+}
+
+.btn-secondary {
+  width: 100%;
+  max-width: 133px;
+  height: 40px;
+  border-radius: 8px;
+  background-color: var(--light-gray-3);
+  color: var(--primary-color);
+  font-size: 0.85rem;
+  font-weight: 700;
+  letter-spacing: 0.75px;
+  line-height: 22px;
+  margin-top: 20px;
+  cursor: pointer;
+  transition: all 0.3s ease-in-out;
+  border: none;
+}
+
+.btn-secondary:hover {
+  background-color: var(--secondary-color);
+  color: var(--light-gray-4);
+}
+
+.btn-shop {
+  width: 100%;
+  max-width: 778px;
+  height: 50px;
+  border-radius: 8px;
+  background-color: var(--warning-color);
+  border: none;
+  cursor: pointer;
+}
