+++ src/components/Buttons/index.jsx
@@ -1,15 +1,22 @@
-import { ButtonIcon, ButtonPrimary, ButtonSecundary } from './styles';
+import { 
+  ButtonIcon, 
+  ButtonPrimary, 
+  ButtonSecundary, 
+  ButtonShop,
+  PrimaryBtn,
+  SecundaryBtn,
+  IconBtn,
+  ShopBtn
+} from './ButtonComponents';
 
-export const PrimaryBtn = ({ children }) => {
-  return <ButtonPrimary>{children}</ButtonPrimary>;
-};
-export const SecundaryBtn = () => {
-  return <ButtonSecundary>{children}</ButtonSecundary>;
-};
-export const IconBtn = () => {
-  return <ButtonIcon>{children}</ButtonIcon>;
-};
-
-export const ShopBtn = ({ children }) => {
-  return <ButtonShop>{children}</ButtonShop>;
+// Re-exporta todos os componentes
+export {
+  ButtonIcon,
+  ButtonPrimary,
+  ButtonSecundary,
+  ButtonShop,
+  PrimaryBtn,
+  SecundaryBtn,
+  IconBtn,
+  ShopBtn
 };
