+++ src/components/Logo/index.jsx
@@ -2,11 +2,11 @@ import IconLogo from '../../assets/images/Vector.svg';
 import ImageLogo from '../../assets/images/DigitalStore.svg';
 import IconLogo2 from '../../assets/images/Vector-white.svg';
 import ImageLogo2 from '../../assets/images/DigitalStore-white.svg';
-import { ContainerLogo } from './styles';
+import './Logo.css';
 
 const Logo = ({ $footer }) => {
   return (
-    <ContainerLogo>
+    <div className="logo-container">
       <div className='content-icon'>
         {$footer ? (
           <img src={IconLogo2} alt='Logo Icon' className='icon-logo' />
@@ -16,12 +16,12 @@ const Logo = ({ $footer }) => {
       </div>
       <div className='content-image'>
         {$footer ? (
-          <img src={ImageLogo2} alt='Logo Image' className='image-logo' />
+          <img src={ImageLogo2} alt='Logo Image' className='logo-image' />
         ) : (
-          <img src={ImageLogo} alt='Logo Image' className='image-logo' />
+          <img src={ImageLogo} alt='Logo Image' className='logo-image' />
         )}
       </div>
-    </ContainerLogo>
+    </div>
   );
 };
