+++ src/components/ProductCard/ProductCard.css
@@ -1,0 +1,188 @@
+/* ProductCard Component Styles */
+
+.product-card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 8px;
+  width: 100%;
+  max-width: 292px;
+  height: 100%;
+  max-height: 439px;
+  cursor: pointer;
+  transition: transform 0.3s ease;
+}
+
+.product-card:hover {
+  transform: translateY(-5px);
+}
+
+.product-card .card-product {
+  width: 292px;
+  height: 321px;
+  flex-shrink: 0;
+  background-color: var(--white);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 4px;
+  overflow: hidden;
+  padding: 38px 20px 0;
+  position: relative;
+}
+
+.product-card .card-product .discount {
+  position: absolute;
+  top: 20px;
+  left: 20px;
+  background-color: var(--card-tag-color);
+  color: var(--dark-gray-2);
+  padding: 5px 15px;
+  border-radius: 29px;
+  font-size: 0.9rem;
+  font-weight: 700;
+  letter-spacing: 0.75px;
+  text-transform: uppercase;
+}
+
+.product-card .card-product img {
+  width: 248px;
+  height: 134px;
+  display: block;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%) rotate(-30deg);
+}
+
+.product-card p {
+  font-size: 0.7rem;
+  font-weight: 700;
+  color: var(--light-gray);
+  line-height: 24px;
+  letter-spacing: 0.75px;
+}
+
+.product-card h4 {
+  font-size: 1.5rem;
+  font-weight: 400;
+  line-height: 24px;
+  letter-spacing: 0.75px;
+  color: var(--dark-gray-2);
+}
+
+.product-card .card-info {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+}
+
+.product-card .card-info .original-price {
+  color: var(--light-gray);
+  font-feature-settings:
+    'liga' off,
+    'clig' off;
+  font-family: Inter;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 38px;
+  letter-spacing: 0.75px;
+  text-decoration: line-through;
+}
+
+.product-card .card-info .discount-price {
+  color: var(--dark-gray);
+  font-feature-settings:
+    'liga' off,
+    'clig' off;
+  font-family: Inter;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 38px;
+  letter-spacing: 0.75px;
+}
+
+/* Responsividade */
+@media (max-width: 1400px) {
+  .product-card {
+    max-width: 220px;
+    height: auto;
+  }
+
+  .product-card .card-product {
+    height: 265px;
+  }
+
+  .product-card .card-product img {
+    width: 190px;
+  }
+
+  .product-card h4 {
+    font-size: 1.1rem;
+  }
+
+  .product-card .card-info .original-price {
+    font-size: 1.3rem;
+    line-height: 28px;
+    letter-spacing: -0.75px;
+  }
+
+  .product-card .card-info .discount-price {
+    font-size: 1.3rem;
+    line-height: 28px;
+    letter-spacing: -0.75px;
+  }
+}
+
+@media (max-width: 460px) {
+  .product-card {
+    width: 100%;
+    min-width: 163px;
+    height: 290px;
+    align-items: flex-start;
+  }
+
+  .product-card a {
+    width: 100%;
+    max-width: 163px;
+  }
+
+  .product-card .card-product {
+    width: 100%;
+    padding: 0px;
+    height: 179px;
+  }
+
+  .product-card .card-product .discount {
+    left: 30px;
+  }
+
+  .product-card .card-product img {
+    width: 120px;
+    top: 40px;
+    left: 14px;
+    transform: rotate(-30deg);
+  }
+
+  .product-card h4 {
+    font-size: 0.8rem;
+  }
+
+  .product-card .card-info {
+    gap: 8px;
+  }
+
+  .product-card .card-info .original-price {
+    font-size: 0.8rem;
+    line-height: 20px;
+    letter-spacing: -0.75px;
+  }
+
+  .product-card .card-info .discount-price {
+    font-size: 0.8rem;
+    line-height: 20px;
+    letter-spacing: -0.75px;
+  }
+}
