+++ src/components/ProductListing/ProductListing.css
@@ -1,0 +1,22 @@
+/* ProductListing Component Styles */
+
+.product-listing-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 24px;
+  width: 100%;
+  justify-content: center;
+}
+
+@media (max-width: 1400px) {
+  .product-listing-container {
+    gap: 20px;
+  }
+}
+
+@media (max-width: 460px) {
+  .product-listing-container {
+    gap: 15px;
+    justify-content: space-around;
+  }
+}
