+++ src/components/Section/index.jsx
@@ -1,34 +1,32 @@
-import {
-  SectionContent,
-  SectionHeader,
-  SectionLink,
-  SectionTitle,
-  StyledSection,
-} from './styles';
+import './Section.css';
 import { FaArrowRightLong } from 'react-icons/fa6';
 
-const Section = ({ title, titleAlign = 'left', link, children }) => {
+const Section = ({ title, titleAlign = 'left', link, children, $paddingBottom, className = '' }) => {
   console.log(link);
   // Determina o alinhamento do título padrão 'left' se não for fornecido ou for diferente de 'center'
   const alignment = titleAlign === 'center' ? 'center' : 'left';
+  
+  const sectionClasses = `section ${$paddingBottom ? 'padding-bottom' : ''} ${className}`;
+  const headerClasses = `section-header ${alignment === 'center' ? 'center' : ''}`;
+  
   return (
-    <StyledSection>
-      <SectionHeader $align={alignment}>
-        {title && <SectionTitle $align={alignment}>{title}</SectionTitle>}
+    <section className={sectionClasses}>
+      <div className={headerClasses}>
+        {title && <h2 className="section-title">{title}</h2>}
         {link &&
           link.text &&
           link.href && ( // Renderiza o link apenas se ele existir e tiver texto e href
-            <SectionLink href={link.href}>
+            <a href={link.href} className="section-link">
               {link.text} <FaArrowRightLong size={15} />
-            </SectionLink>
+            </a>
           )}
-      </SectionHeader>
+      </div>
 
-      <SectionContent>
+      <div className="section-content">
         {children}{' '}
         {/* Aqui o conteúdo passado entre as tags <Section> será renderizado */}
-      </SectionContent>
-    </StyledSection>
+      </div>
+    </section>
   );
 };
