+++ src/data/products.json
@@ -1,0 +1,262 @@
+[
+  {
+    "id": 1,
+    "name": "iPhone 15 Pro Max 256GB",
+    "description": "iPhone 15 Pro Max com chip A17 Pro, câmera de 48MP e tela Super Retina XDR de 6,7 polegadas",
+    "price": 8999.99,
+    "priceDiscount": 7999.99,
+    "category": "smartphone",
+    "brand": "Apple",
+    "image": "/images/iphone-15-pro-max.jpg",
+    "rating": 4.8,
+    "inStock": true,
+    "tagValue": "11% Off"
+  },
+  {
+    "id": 2,
+    "name": "Samsung Galaxy S24 Ultra 512GB",
+    "description": "Galaxy S24 Ultra com S Pen integrada, câmera de 200MP e tela Dynamic AMOLED 2X de 6,8 polegadas",
+    "price": 7999.99,
+    "priceDiscount": 6999.99,
+    "category": "smartphone",
+    "brand": "Samsung",
+    "image": "/images/galaxy-s24-ultra.jpg",
+    "rating": 4.7,
+    "inStock": true,
+    "tagValue": "13% Off"
+  },
+  {
+    "id": 3,
+    "name": "MacBook Air M3 13\" 256GB",
+    "description": "MacBook Air com chip M3, 8GB de RAM unificada e SSD de 256GB. Tela Liquid Retina de 13,6 polegadas",
+    "price": 9999.99,
+    "priceDiscount": 8799.99,
+    "category": "notebook",
+    "brand": "Apple",
+    "image": "/images/macbook-air-m3.jpg",
+    "rating": 4.9,
+    "inStock": true,
+    "tagValue": "12% Off"
+  },
+  {
+    "id": 4,
+    "name": "Dell XPS 13 Plus Intel i7",
+    "description": "Notebook Dell XPS 13 Plus com processador Intel Core i7-13700H, 16GB RAM e SSD 512GB",
+    "price": 8499.99,
+    "priceDiscount": 7299.99,
+    "category": "notebook",
+    "brand": "Dell",
+    "image": "/images/dell-xps-13-plus.jpg",
+    "rating": 4.6,
+    "inStock": true,
+    "tagValue": "14% Off"
+  },
+  {
+    "id": 5,
+    "name": "Sony WH-1000XM5",
+    "description": "Fone de ouvido wireless com cancelamento de ruído ativo, bateria de 30 horas",
+    "price": 1799.99,
+    "priceDiscount": 1499.99,
+    "category": "headphone",
+    "brand": "Sony",
+    "image": "/images/sony-wh1000xm5.jpg",
+    "rating": 4.8,
+    "inStock": true,
+    "tagValue": "17% Off"
+  },
+  {
+    "id": 6,
+    "name": "AirPods Pro 2ª Geração",
+    "description": "AirPods Pro com cancelamento ativo de ruído, áudio espacial personalizado e estojo MagSafe",
+    "price": 1999.99,
+    "priceDiscount": 1699.99,
+    "category": "headphone",
+    "brand": "Apple",
+    "image": "/images/airpods-pro-2.jpg",
+    "rating": 4.7,
+    "inStock": true,
+    "tagValue": "15% Off"
+  },
+  {
+    "id": 7,
+    "name": "iPad Air 5ª Geração 256GB",
+    "description": "iPad Air com chip M1, tela Liquid Retina de 10,9 polegadas e compatibilidade com Apple Pencil",
+    "price": 4999.99,
+    "priceDiscount": 4299.99,
+    "category": "tablet",
+    "brand": "Apple",
+    "image": "/images/ipad-air-5.jpg",
+    "rating": 4.8,
+    "inStock": true,
+    "tagValue": "14% Off"
+  },
+  {
+    "id": 8,
+    "name": "Samsung Galaxy Tab S9 Ultra",
+    "description": "Tablet Samsung com tela AMOLED de 14,6 polegadas, S Pen incluída e 256GB de armazenamento",
+    "price": 6999.99,
+    "priceDiscount": 5999.99,
+    "category": "tablet",
+    "brand": "Samsung",
+    "image": "/images/galaxy-tab-s9-ultra.jpg",
+    "rating": 4.6,
+    "inStock": true,
+    "tagValue": "14% Off"
+  },
+  {
+    "id": 9,
+    "name": "Samsung Neo QLED 65\" 4K",
+    "description": "Smart TV Samsung Neo QLED 65 polegadas com tecnologia Quantum Matrix e processador Neo Quantum 4K",
+    "price": 7999.99,
+    "priceDiscount": 6799.99,
+    "category": "smart-tv",
+    "brand": "Samsung",
+    "image": "/images/samsung-neo-qled-65.jpg",
+    "rating": 4.7,
+    "inStock": true,
+    "tagValue": "15% Off"
+  },
+  {
+    "id": 10,
+    "name": "LG OLED C3 55\" 4K",
+    "description": "Smart TV LG OLED C3 55 polegadas com processador α9 Gen6 AI e Dolby Vision IQ",
+    "price": 5999.99,
+    "priceDiscount": 4999.99,
+    "category": "smart-tv",
+    "brand": "LG",
+    "image": "/images/lg-oled-c3-55.jpg",
+    "rating": 4.8,
+    "inStock": true,
+    "tagValue": "17% Off"
+  },
+  {
+    "id": 11,
+    "name": "Apple Watch Series 9 45mm",
+    "description": "Apple Watch Series 9 com GPS, tela Always-On Retina e chip S9 SiP",
+    "price": 3499.99,
+    "priceDiscount": 2999.99,
+    "category": "gadget",
+    "brand": "Apple",
+    "image": "/images/apple-watch-series-9.jpg",
+    "rating": 4.7,
+    "inStock": true,
+    "tagValue": "14% Off"
+  },
+  {
+    "id": 12,
+    "name": "Nintendo Switch OLED",
+    "description": "Console Nintendo Switch com tela OLED de 7 polegadas, 64GB de armazenamento interno",
+    "price": 2199.99,
+    "priceDiscount": 1899.99,
+    "category": "gadget",
+    "brand": "Nintendo",
+    "image": "/images/nintendo-switch-oled.jpg",
+    "rating": 4.8,
+    "inStock": true,
+    "tagValue": "14% Off"
+  },
+  {
+    "id": 13,
+    "name": "Xiaomi 13 Pro 256GB",
+    "description": "Smartphone Xiaomi 13 Pro com câmera Leica de 50.3MP, tela AMOLED de 6,73 polegadas",
+    "price": 4999.99,
+    "priceDiscount": 4199.99,
+    "category": "smartphone",
+    "brand": "Xiaomi",
+    "image": "/images/xiaomi-13-pro.jpg",
+    "rating": 4.6,
+    "inStock": true,
+    "tagValue": "16% Off"
+  },
+  {
+    "id": 14,
+    "name": "Google Pixel 8 Pro 128GB",
+    "description": "Google Pixel 8 Pro com chip Tensor G3, câmera de 50MP e recursos de IA avançados",
+    "price": 5999.99,
+    "priceDiscount": 5199.99,
+    "category": "smartphone",
+    "brand": "Google",
+    "image": "/images/google-pixel-8-pro.jpg",
+    "rating": 4.5,
+    "inStock": true,
+    "tagValue": "13% Off"
+  },
+  {
+    "id": 15,
+    "name": "Lenovo ThinkPad X1 Carbon Gen 11",
+    "description": "Notebook empresarial Lenovo ThinkPad com Intel Core i7-1365U, 16GB RAM e SSD 512GB",
+    "price": 9999.99,
+    "priceDiscount": 8499.99,
+    "category": "notebook",
+    "brand": "Lenovo",
+    "image": "/images/thinkpad-x1-carbon-gen11.jpg",
+    "rating": 4.7,
+    "inStock": true,
+    "tagValue": "15% Off"
+  },
+  {
+    "id": 16,
+    "name": "ASUS ROG Zephyrus G14",
+    "description": "Notebook gamer ASUS com AMD Ryzen 9 7940HS, RTX 4060, 16GB RAM e SSD 1TB",
+    "price": 12999.99,
+    "priceDiscount": 10999.99,
+    "category": "notebook",
+    "brand": "ASUS",
+    "image": "/images/asus-rog-zephyrus-g14.jpg",
+    "rating": 4.6,
+    "inStock": true,
+    "tagValue": "15% Off"
+  },
+  {
+    "id": 17,
+    "name": "Bose QuietComfort 45",
+    "description": "Fone de ouvido Bose com cancelamento de ruído ativo, bateria de 24 horas",
+    "price": 1999.99,
+    "priceDiscount": 1599.99,
+    "category": "headphone",
+    "brand": "Bose",
+    "image": "/images/bose-qc45.jpg",
+    "rating": 4.7,
+    "inStock": true,
+    "tagValue": "20% Off"
+  },
+  {
+    "id": 18,
+    "name": "JBL Tune 760NC",
+    "description": "Fone de ouvido JBL com cancelamento ativo de ruído e 35 horas de bateria",
+    "price": 699.99,
+    "priceDiscount": 549.99,
+    "category": "headphone",
+    "brand": "JBL",
+    "image": "/images/jbl-tune-760nc.jpg",
+    "rating": 4.4,
+    "inStock": true,
+    "tagValue": "21% Off"
+  },
+  {
+    "id": 19,
+    "name": "Microsoft Surface Pro 9",
+    "description": "Tablet 2-em-1 Microsoft Surface Pro 9 com processador Intel Core i7, 16GB RAM",
+    "price": 8999.99,
+    "priceDiscount": 7699.99,
+    "category": "tablet",
+    "brand": "Microsoft",
+    "image": "/images/surface-pro-9.jpg",
+    "rating": 4.5,
+    "inStock": true,
+    "tagValue": "14% Off"
+  },
+  {
+    "id": 20,
+    "name": "Lenovo Tab P11 Plus",
+    "description": "Tablet Lenovo com tela 2K de 11 polegadas, processador MediaTek Helio G90T",
+    "price": 1999.99,
+    "priceDiscount": 1599.99,
+    "category": "tablet",
+    "brand": "Lenovo",
+    "image": "/images/lenovo-tab-p11-plus.jpg",
+    "rating": 4.3,
+    "inStock": true,
+    "tagValue": "20% Off"
+  }
+]
