+++ src/main.jsx
@@ -2,14 +2,13 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App.jsx';
-import { GlobalStyle } from './style/GlobalStyle.js';
+import './style/global.css';
 import { CartProvider } from './contexts/CartContext.jsx';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <BrowserRouter>
       <CartProvider>
-        <GlobalStyle />
         <App />
       </CartProvider>
     </BrowserRouter>
