+++ src/pages/Layout/Layout.css
@@ -1,0 +1,11 @@
+/* Layout Styles */
+
+.layout-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.layout-content {
+  flex-grow: 1;
+}
