--- src/pages/Layout/styles.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-`;
-
-export const Content = styled.main`
-  flex-grow: 1;
-`;
