+++ src/pages/LoginPage/index.jsx
@@ -4,7 +4,7 @@ import './styles.css';
 import pexels from '../../assets/images/pexels.svg';
 import pexels2 from '../../assets/images/pexels2.svg';
 import { InputDefault } from '../../components/Input';
-import { ButtonPrimary } from '../../components/Buttons/styles';
+import { ButtonPrimary } from '../../components/Buttons/ButtonComponents';
 
 // Componente principal da página de login
 const LoginPage = () => {
